# refactor(adapters): copilot improve token logic and separate