… lua/codecompanion/adapters/http/copilot/get_models.lua
… tests/adapters/http/copilot/test_helpers.lua
… tests/adapters/http/copilot/test_stats.lua