## Description

The Copilot adapter continues to grow in scope. This refactor separates it into modules and also improves on the token resolution.

This model also makes asynchronous calls to the copilot models endpoint. Taking some benchmarks from inside _copilot/init.lua_:

**Before** - For 5 executions:

470.308459ms
385.574959ms
374.258708ms
387.513334ms
377.740333ms

**After** - For 5 executions:

1.015208ms
0.949083ms
0.986709ms
0.970125ms
1.101834ms

So the chat buffer should feel much snappier.

## Checklist

- [x] I've read the [contributing](https://github.com/olimorris/codecompanion.nvim/blob/main/CONTRIBUTING.md) guidelines and have adhered to them in this PR
- [ ] I've updated `CodeCompanion.has` in the [init.lua](https://github.com/olimorris/codecompanion.nvim/blob/main/lua/codecompanion/init.lua#L239) file for my new feature
- [ ] I've added [test](https://github.com/olimorris/codecompanion.nvim/blob/main/CONTRIBUTING.md#testing) coverage for this fix/feature
- [ ] I've updated the README and/or relevant docs pages
- [ ] I've run `make all` to ensure docs are generated, tests pass and my formatting is applied