@olimorris 2025-08-29
@copilot any views on this refactor?